login/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
docopt==0.6.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
pipreqs==0.4.10
PyMySQL==0.10.1
requests==2.25.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
yarg==0.1.9
